refactor(receipt): [Issue #100-2] receiptRepository を read/write/includes に分割
include 形状を receiptIncludes.ts に集約し、読み取り・統計・書き込みを責務別
ファイルへ分離。既存 import パスは receiptRepository.ts の re-export で維持。
